once_cell

Module imp

Modules

strict 🔒

Structs

Guard 🔒

Drains and notifies the queue of waiters on drop.

OnceCell 🔒

Waiter 🔒

Representation of a node in the linked list of waiters in the RUNNING state. A waiters is stored on the stack of the waiting threads.

Constants

COMPLETE 🔒

COMPLETE_PTR 🔒

INCOMPLETE 🔒

INCOMPLETE_PTR 🔒

RUNNING 🔒

STATE_MASK 🔒

Functions

initialize_or_wait 🔒

wait 🔒