Struct once_cell::race::OnceBool

pub struct OnceBool { /* private fields */ }

A thread-safe cell which can be written to only once.

Implementations

impl OnceBool

pub const fn new() -> OnceBool

Creates a new empty cell.

pub fn get(&self) -> Option<bool>

Gets the underlying value.

pub fn set(&self, value: bool) -> Result<(), ()>

Sets the contents of this cell to value.

Returns Ok(()) if the cell was empty and Err(()) if it was full.

pub fn get_or_init<F>(&self, f: F) -> bool
where F: FnOnce() -> bool,

Gets the contents of the cell, initializing it with f if the cell was empty.

If several threads concurrently run get_or_init, more than one f can be called. However, all threads will return the same value, produced by some f.

pub fn get_or_try_init<F, E>(&self, f: F) -> Result<bool, E>
where F: FnOnce() -> Result<bool, E>,

Gets the contents of the cell, initializing it with f if the cell was empty. If the cell was empty and f failed, an error is returned.

If several threads concurrently run get_or_init, more than one f can be called. However, all threads will return the same value, produced by some f.