once_cell::race
Struct OnceRef

pub struct OnceRef<'a, T> { /* private fields */ }

A thread-safe cell which can be written to only once.

Implementations

impl<'a, T> OnceRef<'a, T>

pub const fn new() -> OnceRef<'a, T>

Creates a new empty cell.

pub fn get(&self) -> Option<&'a T>

Gets a reference to the underlying value.

pub fn set(&self, value: &'a T) -> Result<(), ()>

Sets the contents of this cell to value.

Returns Ok(()) if the cell was empty and Err(value) if it was full.

pub fn get_or_init<F>(&self, f: F) -> &'a T
where F: FnOnce() -> &'a T,

Gets the contents of the cell, initializing it with f if the cell was empty.

If several threads concurrently run get_or_init, more than one f can be called. However, all threads will return the same value, produced by some f.

pub fn get_or_try_init<F, E>(&self, f: F) -> Result<&'a T, E>
where F: FnOnce() -> Result<&'a T, E>,

Gets the contents of the cell, initializing it with f if the cell was empty. If the cell was empty and f failed, an error is returned.

If several threads concurrently run get_or_init, more than one f can be called. However, all threads will return the same value, produced by some f.