Struct once_cell::unsync::Lazy

pub struct Lazy<T, F = fn() -> T> { /* private fields */ }

A value which is initialized on the first access.

Example

use once_cell::unsync::Lazy;

let lazy: Lazy<i32> = Lazy::new(|| {
    println!("initializing");
    92
});
println!("ready");
println!("{}", *lazy);
println!("{}", *lazy);

// Prints:
//   ready
//   initializing
//   92
//   92

Implementations

impl<T, F> Lazy<T, F>

pub const fn new(init: F) -> Lazy<T, F>

Creates a new lazy value with the given initializing function.

Example

use once_cell::unsync::Lazy;

let hello = "Hello, World!".to_string();

let lazy = Lazy::new(|| hello.to_uppercase());

assert_eq!(&*lazy, "HELLO, WORLD!");

pub fn into_value(this: Lazy<T, F>) -> Result<T, F>

Consumes this Lazy returning the stored value.

Returns Ok(value) if Lazy is initialized and Err(f) otherwise.

impl<T, F: FnOnce() -> T> Lazy<T, F>

pub fn force(this: &Lazy<T, F>) -> &T

Forces the evaluation of this lazy value and returns a reference to the result.

This is equivalent to the Deref impl, but is explicit.

Example

use once_cell::unsync::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force(&lazy), &92);
assert_eq!(&*lazy, &92);

pub fn force_mut(this: &mut Lazy<T, F>) -> &mut T

Forces the evaluation of this lazy value and returns a mutable reference to the result.

This is equivalent to the DerefMut impl, but is explicit.

Example

use once_cell::unsync::Lazy;

let mut lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force_mut(&mut lazy), &92);
assert_eq!(*lazy, 92);

pub fn get(this: &Lazy<T, F>) -> Option<&T>

Gets the reference to the result of this lazy value if it was initialized, otherwise returns None.

Example

use once_cell::unsync::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::get(&lazy), None);
assert_eq!(&*lazy, &92);
assert_eq!(Lazy::get(&lazy), Some(&92));

pub fn get_mut(this: &mut Lazy<T, F>) -> Option<&mut T>

Gets the mutable reference to the result of this lazy value if it was initialized, otherwise returns None.

Example

use once_cell::unsync::Lazy;

let mut lazy = Lazy::new(|| 92);

assert_eq!(Lazy::get_mut(&mut lazy), None);
assert_eq!(*lazy, 92);
assert_eq!(Lazy::get_mut(&mut lazy), Some(&mut 92));