Module once_cell::race

Thread-safe, non-blocking, “first one wins” flavor of OnceCell.

If two threads race to initialize a type from the race module, they don’t block, execute initialization function together, but only one of them stores the result.

This module does not require std feature.

Structs

OnceBool

A thread-safe cell which can be written to only once.

OnceBox

A thread-safe cell which can be written to only once.

OnceNonZeroUsize

A thread-safe cell which can be written to only once.