Struct once_cell::sync::Lazy

pub struct Lazy<T, F = fn() -> T> { /* private fields */ }

A value which is initialized on the first access.

This type is thread-safe and can be used in statics.

Example

use std::collections::HashMap;

use once_cell::sync::Lazy;

static HASHMAP: Lazy<HashMap<i32, String>> = Lazy::new(|| {
    println!("initializing");
    let mut m = HashMap::new();
    m.insert(13, "Spica".to_string());
    m.insert(74, "Hoyten".to_string());
    m
});

fn main() {
    println!("ready");
    std::thread::spawn(|| {
        println!("{:?}", HASHMAP.get(&13));
    }).join().unwrap();
    println!("{:?}", HASHMAP.get(&74));

    // Prints:
    //   ready
    //   initializing
    //   Some("Spica")
    //   Some("Hoyten")
}

Implementations

impl<T, F> Lazy<T, F>

pub const fn new(f: F) -> Lazy<T, F>

Creates a new lazy value with the given initializing function.

pub fn into_value(this: Lazy<T, F>) -> Result<T, F>

Consumes this Lazy returning the stored value.

Returns Ok(value) if Lazy is initialized and Err(f) otherwise.

impl<T, F: FnOnce() -> T> Lazy<T, F>

pub fn force(this: &Lazy<T, F>) -> &T

Forces the evaluation of this lazy value and returns a reference to the result. This is equivalent to the Deref impl, but is explicit.

Example

use once_cell::sync::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force(&lazy), &92);
assert_eq!(&*lazy, &92);

pub fn force_mut(this: &mut Lazy<T, F>) -> &mut T

Forces the evaluation of this lazy value and returns a mutable reference to the result. This is equivalent to the Deref impl, but is explicit.

Example

use once_cell::sync::Lazy;

let mut lazy = Lazy::new(|| 92);

assert_eq!(Lazy::force_mut(&mut lazy), &mut 92);

pub fn get(this: &Lazy<T, F>) -> Option<&T>

Gets the reference to the result of this lazy value if it was initialized, otherwise returns None.

Example

use once_cell::sync::Lazy;

let lazy = Lazy::new(|| 92);

assert_eq!(Lazy::get(&lazy), None);
assert_eq!(&*lazy, &92);
assert_eq!(Lazy::get(&lazy), Some(&92));

pub fn get_mut(this: &mut Lazy<T, F>) -> Option<&mut T>

Gets the reference to the result of this lazy value if it was initialized, otherwise returns None.

Example

use once_cell::sync::Lazy;

let mut lazy = Lazy::new(|| 92);

assert_eq!(Lazy::get_mut(&mut lazy), None);
assert_eq!(&*lazy, &92);
assert_eq!(Lazy::get_mut(&mut lazy), Some(&mut 92));